refactor(xapiri): extract updateDepsTab + renderDepsTab → tab_deps.go
Mechanical move per ADR 0014. No behaviour change.
